Dockerizing the application
